feat: add author validation for updat and delete comment

+++ src/services/comments.services.js
@@ -1,4 +1,5 @@
 import { NotFoundException } from '../err/notFoundException.js';
+import { UnauthorizedException } from '../err/unauthorizedException.js';
 import { commentsRepo } from '../repository/comments.repo.js';
 import { worksRepo } from '../repository/works.repo.js';
 
@@ -29,17 +30,27 @@ async function commentExistence(commentId) {
   if (!comment) {
     throw new NotFoundException('댓글을 찾을 수 없습니다.');
   }
-  return !!comment;
+  return comment;
 }
 
-async function updateComment({ commentId, content }) {
-  await commentExistence(commentId);
+async function updateComment({ commentId, userId, content }) {
+  const comment = await commentExistence(commentId);
+  if (userId !== comment.authorId) {
+    throw new UnauthorizedException(
+      '댓글의 작성자가 아니기 때문에 수정할 수 없습니다.',
+    );
+  }
   const data = { content };
   return commentsRepo.updateComment({ commentId, data });
 }
 
-async function deleteComment(commentId) {
-  await commentExistence(commentId);
+async function deleteComment({ commentId, userId }) {
+  const comment = await commentExistence(commentId);
+  if (userId !== comment.authorId) {
+    throw new UnauthorizedException(
+      '댓글의 작성자가 아니기 때문에 삭제할 수 없습니다.',
+    );
+  }
   return commentsRepo.deleteComment(commentId);
 }
